requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apify_client==1.9.2
apify_shared==1.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.11
groq==0.19.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
more-itertools==10.6.0
oauthlib==3.2.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
starlette==0.46.1
tweepy==4.15.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
pytest==7.4.2
pytest-asyncio==0.21.1